Re-organized file system.
