+++ bin/app.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+const index = require("../src/index.js");
+index.main();
